integrations/erc20: new (#139)
I don't use the ABI package to decode these logs because the event 
only has a single item in the ABI encoded field and it is a statically 
sized uint256. Therefore, we can save some compute cycles by just 
converting it directly. 'from' and 'to' are parsed directly from the topics 
array and they are resized to 20 bytes.