shovel: handle unsynchronized node providers (#250)
Background:

When you issue an eth_getLogs request for a block that is
head of a node, the node will respond with an empty array. Eg

	curl https://XXX \
	--header 'Content-Type: application/json' \
	--data '
	{
		"id":"1",
		"jsonrpc":"2.0",
		"method":"eth_getLogs",
		"params":[
			{
				"fromBlock":"0xffffff",
				"toBlock":"0xffffff",
				"address":null,
				"topics":[]
			}
		]
	}
	'

	{"jsonrpc":"2.0","id":"1","result":[]}

Furthermore, hosted node providers will often load balance requests
between nodes. This means that it is possible for a node API to provide
an inconsistent view of the blockchain between requests.

Problem:

Shovel tracks the head of the chain by asking an API provider for its
latest block. If the node provider responds with a block that is ahead
of Shovel then Shovel will begin to index the new block. For
integrations that need logs, Shovel will immediately ask for the logs of
the new block. In some rare cases, this eth_getLogs request will be
routed to a node that doesn't have the latest block. Shovel will
incorrectly assume that the block has no logs and move forward with
indexing.

To solve this problem, we will always issue an eth_getBlockByNumber
request along with the eth_getLogs via the JSON RPC batching mechanism.
If the node that receives the request for logs doesn't have the block,
then we can detect this and return an error thus causing Shovel to try
again.

It may be possible for a node provider to split up batch requests among
different nodes. In this case this fix doesn't really solve the problem.
I did check that the OP proxyd project, when rpc_method_mappings group
both eth_getLogs and eth_getBlockByNumber together, and when
max_upstream_batch_size is <= 2, will keep the batched requests
together. Which is nice.

Here is the proxyd batching code:

https://tinyurl.com/k3nsx8s9
Location:  github.com/ethereum-optimism/optimism/proxyd/server.go#L415