indexsupply.com/shove/docs: rewrite (#252)
The new version of the docs has Overview and Reference sections. I was
always frusterated that there wasn't a canonical description of the
Config object. Now that exists as Reference: Config Object. But I felt
that even with a Config Object Reference, there needs to be some high
level description of the Config Object. So that exists as:
Overview: Config.

I also added a few other things.